HTMX! Create really basic homepage for server
The server homepage can display device information and the
latest system data in a very simple manner. This is good enough
to start working on the client side so I can get a client device
running and publishing real data to store in the db.
Then I will come back to make the server display more useful data.