Simplify devices table, improve server UI
Simplify the devices table to just store an id, name, and
created_at date. This removes the raspberry pi specific columns
I had created. I decided to do this because I want the user to
have to register devices through the server UI using a name for
the device. In the future, a non raspberry pi device might be
used too and this data would not exist for those devices.

Improve the server UI to have ways to register new devices,
get information for a given device, and list all of the
registered devices.